Fixed read-in
Text files had non-traditional spaces between words. I converted these when reading in the text files to fix the analysis. This was realized after meeting with client, and realizing that the word for "I" did not have commonality 1 as would be assumed. Also, a word document with common words listed was created.